+++ R/print.cSEMAssess.R
@@ -216,7 +216,7 @@ print.cSEMAssess <- function(x, ...) {
   
   if(any(names(x) %in% c("Fornell-Larcker", "HTMT"))) {
     cat2("\n\n", rule2("Validity assessment"))
-    if(any(names(x) == "HTMT") && !anyNA(x$HTMT)) {
+    if(any(names(x) == "HTMT") && !is.null(x$HTMT)) {
       cat2("\n\n\tHeterotrait-montrait ratio of correlation matrix (HTMT matrix)\n\n")
       print(x$HTMT)
     }